wip doubly linked list
